Get Invoice Data
This gets the accounting info for an invoice on the SQL Server as a
query.  You assign your form's list record source as the newly created
query.  I also need to confirm the invoice amount hadn't changed from
the original so I create a recordset so I can pull the invoice number.